requirements.txt
# Core dependencies for Pothole Detection System

# Web framework for dashboard
Flask==3.1.3
Werkzeug==3.1.6

# Computer vision and image processing
opencv-python==4.8.0.74

# HTTP requests for Roboflow API
requests==2.31.0

# Geolocation services
geocoder==1.38.1

# Environment variables management (for API keys)
python-dotenv==1.0.0

# JSON parsing
Jinja2==3.1.2
